+++ cfn-resources/integration-tests/trigger/trigger_template.json
@@ -1,0 +1,44 @@
+{
+  "AWSTemplateFormatVersion": "2010-09-09",
+  "Description": "This template creates a Trigger on the MongoDB Cluster, this will be billed to your Atlas account.",
+  "Mappings": {},
+  "Resources": {
+    "EventTrigger": {
+      "Type": "MongoDB::Atlas::Trigger",
+      "Properties": {
+        "Name": "cfn-trigger-example",
+        "Type": "DATABASE",
+        "DatabaseTrigger": {
+          "OperationTypes": [
+            "INSERT"
+          ],
+          "Database": "sample_mflix",
+          "Collection": "comments",
+          "ServiceId": "64020111d61a66df7ea2c732"
+        },
+        "EventProcessors": {
+          "FUNCTION": {
+            "FuncConfig": {
+              "FunctionName": "myFuncInTrigger",
+              "FunctionId": "64020cfc5ef00ef0be9584e4"
+            }
+          }
+        },
+        "AppId": "64020111d61a66df7ea2c72b",
+        "ProjectId": "63f4df9e1c744217893c19f7",
+        "Profile": "default"
+      }
+    }
+  },
+  "Outputs": {
+    "Id": {
+      "Description": "Id of the trigger",
+      "Value": {
+        "Fn::GetAtt": [
+          "EventTrigger",
+          "Id"
+        ]
+      }
+    }
+  }
+}
